SDSTOR-1677: Add `--version` flag to SDS binaries.
By default, all binaries will have a "-V/--version" flag that can be
used to have the process simply log the package name and version to the
logging sinks and exit. This function requires that the following be
defined by the code calling `SetLogger` either by the compiler arg via
CMake or `#define` in the code itself:

// -DPACKAGE_NAME="MyPackage"
// -DPACKAGE_VERSION="1.0.0"

Example of doing this in the build files can be seen in test_package.