SDSTOR-5593 Fix for metrics remain exposed after volume is deleted. (#164)
Fundemental problem is metrics uses prometheus labels to register and
when the volume is deleted, metrics actually deregisters itself.
However, it does not remove the label from prometheus version of metric
family. Hence it continues to emit the metric during prometheus scraping
even though there is no change or update to that metric. Fix is to
actually remove the label corresponding to the volume. This version of
prometheus, however does not allow Unregister of the metric family
itself.

After fixing this, the singleton destruction order has changed, result
in crashes when application starts to use it. This was fixed in develop
branch and some of the fixes are backported.